requirements.txt
django==2.1.7
pytz==2018.9
pyctogram
vk_text_parser
mq_consumer
pillow
python-telegram-bot
telegram_log
sqlparse
psycopg2-binary
requests[socks]
django-pglocks
apache-airflow[celery,crypto]
tenacity>=4.11.0
gunicorn
redis==3.2.1
transliterate
google_images_download
filelock
psutil